Module uu_numfmt::format

Structs

WhitespaceSplitter

Iterate over a line’s fields, where each field is a contiguous sequence of non-whitespace, optionally prefixed with one or more characters of leading whitespace. Fields are returned as tuples of (prefix, field).

Functions

div_round

Divide numerator by denominator, with rounding.

format_and_print

Format a line of text according to the selected options.